### 2024-12-13 #7369 @mbarzegary
Wait! Is there any JOSS paper here in this submission?
@ebknudsen
