### 2024-12-13 #7369 @ebknudsen
I think I have been confused by repos - the paper lives here: https://github.com/ebknudsen/c2omc_joss
I will have to look through the docs again to see what I need to do in order to make it available to you.
