+++ paper.md
@@ -1,0 +1,297 @@
+---
+title: "CAD_to_OpenMC: from CAD design to particle transport"
+tags:
+    - Python
+    - Monte Carlo
+    - Model preparation
+    - CAD
+    - Mesh
+
+authors:
+    - name: Erik Bergbäck Knudsen
+      orcid: 0000-0003-1827-9188
+      corresponding: true
+      affiliation: 1
+      equal contribution: true
+    - name: Lorenzo Chierici
+      affiliation: 2
+
+affiliations:
+    - name: United Neux, Denmark
+      index: 1
+    - name: Copenhagen Atomics A/S, Denmark
+      index: 2
+
+date: XX September 2024
+bibliography: bibliography.bib
+---
+
+# Summary
+<!-- Here's a citation @knudsen2013mcxtrace , or in a different format [@knudsen2013mcxtrace].-->
+We present CAD_to_OpenMC - a practical tool for allowing geometries designed in
+CAD-systems to be used in Monte Carlo particle transport simulation systems.
+Written in python, and easily installable through pip and/or conda channels,
+CAD_to_openMC does so by importing a geometry in the form of a step-file, a
+common export format among many CAD-systems, identifying objects within that
+file, and running a triangularization algorithm on all surfaces contained within
+the geometry.
+The geometry is written to a HDF5-file in a format defined by the
+DAGMC-library, which is usable by the main particle-transport packages.
+
+
+# Statement of need
+Most Monte Carlo neutronics-, and general particle transport (e.g. OpenMC, GEANT4, MCNP, Fluka [@openmc_2013;
+@geant4_2003; @mcnp_2022; @fluka_2005; @fluka_2014]) software packages use the
+concept of Constructive Solid Geometry (CSG) to the describe the geometry or
+scene in which a computations are to performed. Computations generally amount
+to tracing a large number of rays through the geometry. Such models range in
+complexity from simple scenes of a only a few geometrical objects to several
+million, but are generally considered tricky to work with - in that all object
+boundaries must be constructed from analytical descriptions of geometrical
+primaries (cylinders , planes, 2nd order surfaces, etc...).
+Most engineering design, however, is performed using CAD-tools, which are
+optimized for ease of design.
+
+These facts form a compelling argument for bridging the gap between these
+worlds. So far a number of solutions have been proposed which can mostly be put
+two categories.
+
+1. tools that convert CAD-models into CSG [@du_inversecsg_2018; @catalan_geouned_2024], and
+2. tools that allow ray-tracing on CAD-geometries, or close derivatives thereof.
+
+In the second category, one of the most wide-spread open source solutions is DAGMC [@wilson_acceleration_2010],
+a library with couplings to all of OpenMC, MCNP, fluka, and Geant4.
+DAGMC is proven to be fast and reliable, yet one hurdle remains: DAGMC
+requires a discretized (meshed) approximation of the CAD-geometry. In general
+not a trivial requirement.
+
+CAD-software packages are for the most
+part proprietary packages with little to no focus on being interchangeable. One
+upshot of this is a notable lack of overlapping standard formatting between
+packages, the exception being the STEP-format which is backed by an
+international standard [@stepStd_2024]. Thus, a tool should support this format, if
+it is not to be tied to any particular CAD-engine.
+
+The CAD_to_OpenMC tool eases the process of generating a meshed description
+of a CAD-generated geometry (in the form of a step-file) ready for inclusion in
+transport codes through DAGMC.
+
+# Method
+CAD_to_OpenMC uses OCCT [@occt3d] to interact with CAD-geometries
+and its' handling of step-files. Once the geometry has been imported, one of
+several meshing back-ends may be called to create a discretized version of the
[…]
