+++ bibliography.bib
@@ -1,0 +1,1735 @@
+@article{wilson_acceleration_2010,
+	title = {Acceleration techniques for the direct use of {CAD}-based geometry in fusion neutronics analysis},
+	volume = {85},
+	rights = {https://www.elsevier.com/tdm/userlicense/1.0/},
+	issn = {09203796},
+	url = {https://linkinghub.elsevier.com/retrieve/pii/S0920379610002425},
+	doi = {10.1016/j.fusengdes.2010.05.030},
+	pages = {1759--1765},
+	number = {10},
+	journaltitle = {Fusion Engineering and Design},
+	shortjournal = {Fusion Engineering and Design},
+	author = {Wilson, Paul P.H. and Tautges, Timothy J. and Kraftcheck, Jason A. and Smith, Brandon M. and Henderson, Douglass L.},
+	urldate = {2024-09-18},
+	date = {2010-12},
+	langid = {english},
+}
+
+@article{catalan_geouned_2024,
+	title = {{GEOUNED}: A new conversion tool from {CAD} to Monte Carlo geometry},
+	volume = {56},
+	issn = {17385733},
+	url = {https://linkinghub.elsevier.com/retrieve/pii/S1738573324000548},
+	doi = {10.1016/j.net.2024.01.052},
+	shorttitle = {{GEOUNED}},
+	pages = {2404--2411},
+	number = {6},
+	journaltitle = {Nuclear Engineering and Technology},
+	shortjournal = {Nuclear Engineering and Technology},
+	author = {Catalán, J.P. and Sauvan, P. and García, J. and Alguacil, J. and Ogando, F. and Sanz, J.},
+	urldate = {2024-09-18},
+	date = {2024-06},
+	langid = {english},
+}
+
+@article{du_inversecsg_2018,
+	title = {{InverseCSG}: automatic conversion of 3D models to {CSG} trees},
+	volume = {37},
+	issn = {0730-0301, 1557-7368},
+	url = {https://dl.acm.org/doi/10.1145/3272127.3275006},
+	doi = {10.1145/3272127.3275006},
+	shorttitle = {{InverseCSG}},
+	abstract = {While computer-aided design is a major part of many modern manufacturing pipelines, the design files typically generated describe raw geometry. Lost in this representation is the procedure by which these designs were generated. In this paper, we present a method for reverse-engineering the process by which 3D models may have been generated, in the language of constructive solid geometry ({CSG}). Observing that {CSG} is a formal grammar, we formulate this inverse {CSG} problem as a program synthesis problem. Our solution is an algorithm that couples geometric processing with state-of-the-art program synthesis techniques. In this scheme, geometric processing is used to convert the mixed discrete and continuous domain of {CSG} trees to a pure discrete domain where modern program synthesizers excel. We demonstrate the efficiency and scalability of our algorithm on several different examples, including those with over 100 primitive parts. We show that our algorithm is able to find simple programs which are close to the ground truth, and demonstrate our method's applicability in mesh re-editing. Finally, we compare our method to prior state-of-the-art. We demonstrate that our algorithm dominates previous methods in terms of resulting {CSG} compactness and runtime, and can handle far more complex input meshes than any previous method.},
+	pages = {1--16},
+	number = {6},
[…]
