### 2025-01-23 #7369 @bonh
Looks interesting and I'd be happy to review it
