### 2025-01-24 #7710 @editorialbot
[REVIEW]: CAD_to_OpenMC: from CAD design to particle transport
**Submitting author:** @ebknudsen (<a href="http://orcid.org/0000-0003-1827-9188">Erik B Knudsen</a>)
**Repository:** https://www.github.com/openmsr/CAD_to_OpenMC
**Branch with paper.md** (empty if default branch): paper_joss
**Version:** v1.0.0
