### 2025-01-24 #7710 @mbarzegary
@ebknudsen this is where the review takes place. Please keep an eye out for comments here from the reviewers, as well as any issues opened by them on your software repository. I recommend you aim to respond to these as soon as possible, and you can address them straight away as they come in if you like, to ensure we do not loose track of the reviewers.

I suggest that you start by removing the uncited references from your bib file. It will help us later to spot the invalid entries with missing DOIs according to the reference check above, otherwise you may need to fix them all.
