### 2025-01-24 acd7b8f
remove unused bib entries
