+++ bibliography.bib
@@ -50,24 +50,7 @@
 	langid = {english},
 }
 
-@article{hagopian_updating_2018,
-	title = {Updating the {GODIVA}-{IV} benchmark},
-	volume = {119},
-	pages = {833},
-	number = {1},
-	journaltitle = {Trans. Am. Nucl. Soc},
-	author = {Hagopian, Victoria and Goda, Joetta and Hutchinson, Jesson and Kimpland, Robert and {McKenzie}, George},
-	date = {2018},
-}
-
-@report{fasso_fluka_2005,
-	title = {{FLUKA}: a multi-particle transport code},
-	institution = {{CERN}-2005-10},
-	author = {Fasso, Alberto and Ferrari, Alfredo and Ranft, Johannes and Sala, Paola. R.},
-	date = {2005},
-}
-
-@article{agostinelli_geant4simulation_2003,
+@article{geant4_2003,
 	title = {Geant4—a simulation toolkit},
 	volume = {506},
 	rights = {https://www.elsevier.com/tdm/userlicense/1.0/},
@@ -84,7 +67,7 @@
 	langid = {english},
 }
 
-@report{kulesza_mcnp_2022,
+@report{mcnp_2022,
 	location = {Los Alamos, {NM}, {USA}},
 	title = {{MCNP} Code Version 6.3.0 Theory \& User Manual},
 	url = {https://www.osti.gov/biblio/1889957},
@@ -95,7 +78,8 @@
 	doi = {10.2172/1889957},
 }
 
-@article{bohlen_fluka_2014,
+
+@article{fluka_2014,
 	title = {The {FLUKA} Code: Developments and Challenges for High Energy and Medical Applications},
 	volume = {120},
 	issn = {00903752},
@@ -123,1501 +107,6 @@
 	doi = {10.2172/4237975},
 }
 
-@article{romano_openmc_2013,
-	title = {The {OpenMC} Monte Carlo particle transport code},
-	volume = {51},
-	rights = {https://www.elsevier.com/tdm/userlicense/1.0/},
-	issn = {03064549},
-	url = {https://linkinghub.elsevier.com/retrieve/pii/S0306454912003283},
-	doi = {10.1016/j.anucene.2012.06.040},
-	pages = {274--281},
-	journaltitle = {Annals of Nuclear Energy},
-	shortjournal = {Annals of Nuclear Energy},
-	author = {Romano, Paul K. and Forget, Benoit},
-	urldate = {2024-08-12},
-	date = {2013-01},
-	langid = {english},
-}
-
-@article{ewald_zur_1913,
-	title = {Zur Theori der Interferenzen der Röntgenstrahlen in Kristallen},
-	volume = {14},
-	pages = {465--472},
-	number = {11},
-	journaltitle = {Physikalische Zeitschrift},
-	author = {Ewald, P. P.},
-	date = {1913},
-}
-
-@article{bess_current_2020,
-	title = {Current overview of {ICSBEP} and {IRPhEP} benchmark evaluation practices},
-	volume = {239},
-	rights = {https://creativecommons.org/licenses/by/4.0/},
-	issn = {2100-014X},
-	url = {https://www.epj-conferences.org/10.1051/epjconf/202023918002},
-	doi = {10.1051/epjconf/202023918002},
[…]
