### 2025-01-24 #7710 @ebknudsen
@mbarzegary Thanks for the comment - I have now cleaned up the bib file.
