### 2025-01-27 99915bf
test script that installs weekly using micromamba
