### 2025-01-27 598bf74
add more installs
