### 2025-02-17 #7710 @ebknudsen
> [@ebknudsen](https://github.com/ebknudsen) This is a really nice contribution.
>
> A few comments on the paper:
>
>     * line 58, when you discuss surfaces, curves, etc. it may be useful to distinguish that OCCT is using a BREP rather than CSG. And describe how BREP makes use of topology and shape information. (this isn't required, but I don't think most neutronics folks are familiar with how BREP and CSG models are different).
>
>     * You are missing some aspects of the topology in your description, i.e., regions, shells, loops
>
>     * in line 79, the abbreviation C2O needs to be specified
>
>     * Meshing backend: In the documentation, can you give a sense of what's a large model that gmsh would have issues with? i.e., is it 100 regions, 1000 regions, etc.? This is probably too detailed for the paper.
>
>     * line 89, should have comma after "that is"
>
>     * ICSBEP handbook link is broken, gives “withdrawn content error”
>
>     * did you include the STEP files for the 3 reactors discussed in your project repository? I did not find them.
>
>     * line 112,  missing period after “2D-computations”
>
>     * line 119, missing word after “built-in”
>
>     * did you have CSG models of these reactors as well? If so, did you compare the volume difference between CSG model and BREP model?
>
>     * For the volume comparison, consider using the gmsh backend and computing a full 3D mesh from your surface mesh. This will give you a deterministic (and exact) volume for the meshed geometry rather than the stochastic sampling method.
>
>     * line 146, missing word after “order of”
>
>     * line 163, missing words
>
>     * line 166, comma before etc.
>
>
> Note that the reviewers need to be able to reproduce the results presented in the paper. To that end, please provide:
>
>     * STEP files for reactors and scripts used to generate surface triangularizations presented in the paper
>
>     * CSG geometry for results presented
>
>     * OpenMC input files to reproduce tables 1, 2, and 3
>
>     * OpenMC files to compute the volumes with the stochastic sampling method show in figure 3

Thank you very much for the thorough review. We will look at the issues you have raised as soon as all reviews are in.
