### 2025-02-17 openmsr/CAD_to_OpenMC#126 @bonh
JOSS: Contributions of the authors of the paper
Just checking the contributions of the authors to the code and cannot find any commits from Lorenzo Chierici. Usually that doesn't mean anything but please clarify.

@https://github.com/openjournals/joss-reviews/issues/7710
