### 2025-02-17 openmsr/CAD_to_OpenMC#127 @bonh
JOSS: Other software packages, same job?
In the readme you are stating that

> CAD_to_OpenMC is certainly not the only package that tries to solve this problem. Other options include:
>
> [cad_to_dagmc](https://github.com/fusion-energy/cad_to_dagmc)
> [stellarmesher](https://github.com/Thea-Energy/stellarmesh)
> These two packages solve the same problem as CAD_to_OpenMC, and in fact in many cases are interchangeable. See for instance >> [benchmark-zoo] ( https://github.com/fusion-energy/model_benchmark_zoo ).

In the readme from stellarmesh it says:

> Stellarmesh is developed by the original authors of [Thea-Energy/stellarmesh](https://github.com/Thea-Energy/stellarmesh), [CAD-to-DAGMC](https://github.com/fusion-energy/cad_to_dagmc), and [CAD-to-OpenMC](https://github.com/openmsr/CAD_to_OpenMC).

These other projects, stellarmesh and cad_to_dagmc, seem to be active, too. And share developers?

I couldn't find any comments about other software packages in the paper. Perhaps you should clarify what makes your package unique?

https://github.com/openjournals/joss-reviews/issues/7710
