### 2025-02-17 openmsr/CAD_to_OpenMC#126 @ebknudsen
Lorenzo has contributed significantly to the article text not through commits but through other channels.
Also he's been involved in the "paper"-design of the the CAD_to_OpenMC, as well as contributed extensive testing and use of it. Most of those cases are unfortunately company IP.
