### 2025-02-17 openmsr/CAD_to_OpenMC#126 @bonh
thx!
