### 2025-02-17 #7710 @bonh
@jacobmerson

I agree with that point:

> Note that the reviewers need to be able to reproduce the results presented in the paper. To that end, please provide:

> STEP files for reactors and scripts used to generate surface triangularizations presented in the paper
> CSG geometry for results presented
> OpenMC input files to reproduce tables 1, 2, and 3
> OpenMC files to compute the volumes with the stochastic sampling method show in figure 3

Perhaps you'd like open an issue over at the software's repo so that I can comment there? Easier to keep track that way. 😃
