### 2025-02-17 openmsr/CAD_to_OpenMC#128 @bonh
JOSS: Dependencies?
I cannot locate a list of dependencies.

Starting from vanilla conda env required around 340 MB of download. Just want to make sure that this is on purpose: Are all these packages pulled by conda really required?

@https://github.com/openjournals/joss-reviews/issues/7710
