### 2025-02-17 openmsr/CAD_to_OpenMC#128 @bonh
Ah, there is a `requirements.txt` with dependencies. Perhaps it would be helpful to include the list in the readme.
