### 2025-02-17 openmsr/CAD_to_OpenMC#129 @bonh
JOSS: Functionality documentation
I didn't try the script to generate api documentation yet. Perhaps you'd like to generate the docs and host them online? Makes scrolling easier 😄

@https://github.com/openjournals/joss-reviews/issues/7710
