### 2025-02-17 openmsr/CAD_to_OpenMC#130 @bonh
JOSS: Community guidelines
I couldn't find information on: "Are there clear guidelines for third parties wishing to 1. Contribute to the software 2. Report issues or problems with the software 3. Seek support" (copied from JOSS)

@https://github.com/openjournals/joss-reviews/issues/7710
