### 2025-02-17 openmsr/CAD_to_OpenMC#131 @bonh
JOSS: Verify new installation
Perhaps you'd like to add an automated quantitative tests so that I can verify my installation?

Right now I can run through your examples and qualitatively verify the results but a more systematic approach would be better, right?

@https://github.com/openjournals/joss-reviews/issues/7710
