### 2025-02-17 openmsr/CAD_to_OpenMC#132 @bonh
JOSS: paper summary
Matter of style, but I think it would be helpful to extent the summary to include a short explanation for what particle transport code is good for; half a sentence.

https://github.com/openjournals/joss-reviews/issues/7710
