### 2025-02-17 openmsr/CAD_to_OpenMC#133 @jacobmerson
JOSS: reproducibility of results in paper
Note that the reviewers need to be able to reproduce the results presented in the paper. To that end, please provide:

- STEP files for reactors and scripts used to generate surface triangularizations presented in the paper
- CSG geometry for results presented
- OpenMC input files to reproduce tables 1, 2, and 3
- OpenMC files to compute the volumes with the stochastic sampling method show in figure 3

https://github.com/openjournals/joss-reviews/issues/7710#issuecomment-2662376142
