### 2025-02-17 #7710 @jacobmerson
> Perhaps you'd like open an issue over at the software's repo so that I can comment there? Easier to keep track that way. 😃

Good idea see here: https://github.com/openmsr/CAD_to_OpenMC/issues/133#issue-2858507619
