### 2025-02-17 openmsr/CAD_to_OpenMC#133 @ebknudsen
Good points - we agree completely. This will be included.
