### 2025-02-17 openmsr/CAD_to_OpenMC#127 @jacobmerson
I agree, I think there needs to be some more details in the documentation about why one should choose one of these packages over the other. Especially since they share developers. Is it just a matter of style, or are there technical differences, etc.
