### 2025-02-17 openmsr/CAD_to_OpenMC#134 @jacobmerson
JOSS: paper comments
To make the review easier to track, I am adding an issue directly on this repository. The original comment is here:  https://github.com/openjournals/joss-reviews/issues/7710#issuecomment-2662376142

- line 58, when you discuss surfaces, curves, etc. it may be useful to distinguish that OCCT is using a BREP rather than CSG. And describe how BREP makes use of topology and shape information. (this isn't required, but I don't think most neutronics folks are familiar with how BREP and CSG models are different).
- You are missing some aspects of the topology in your description, i.e., regions, shells, loops
- in line 79, the abbreviation C2O needs to be specified
- Meshing backend: In the documentation, can you give a sense of what's a large model that gmsh would have issues with? i.e., is it 100 regions, 1000 regions, etc.? This is probably too detailed for the paper.
- line 89, should have comma after "that is"
- ICSBEP handbook link is broken, gives “withdrawn content error”
- did you include the STEP files for the 3 reactors discussed in your project repository? I did not find them.
- line 112,  missing period after “2D-computations”
- line 119, missing word after “built-in”
- did you have CSG models of these reactors as well? If so, did you compare the volume difference between CSG model and BREP model?
- For the volume comparison, consider using the gmsh backend and computing a full 3D mesh from your surface mesh. This will give you a deterministic (and exact) volume for the meshed geometry rather than the stochastic sampling method.
- line 146, missing word after “order of”
- line 163, missing words
- line 166, comma before etc.
