### 2025-02-18 openmsr/CAD_to_OpenMC#135 @jacobmerson
JOSS: constrain trimesh version
When installing CAD_to_OpenMC via pip, pip decides to download every version of trimesh. To avoid downloading hundreds of different versions, please consider constraining it to a specific version or subset of versions.

https://github.com/openjournals/joss-reviews/issues/7710
