### 2025-02-18 openmsr/CAD_to_OpenMC#136 @jacobmerson
JOSS: example script broken
The example script provided references geometry that does not exist here:

https://github.com/openmsr/CAD_to_OpenMC/blob/b8d07bea6ac055d512d96ec5983f31a9bfce0f15/examples/create_mesh_sample.py#L51

I tried changing the geometry to a couple different options in the `examples/step_files` folder, however I get the following message/error that seems to indicate that the vtk writer is being used instead of the hdf5 writer.

```
Using default writer WriteVtk for file output.h5m
Traceback (most recent call last):
  File "/lore/mersoj2/cad-to-openmc-review/CAD_to_OpenMC/examples/create_mesh_sample.py", line 56, in <module>
    a.solids_to_h5m(backend='stl', h5m_filename='output.h5m')
  File "/lore/mersoj2/cad-to-openmc-review/CAD_to_OpenMC/src/CAD_to_OpenMC/assembly.py", line 533, in solids_to_h5m
    self.stl2h5m(stl_list, h5m_path.name, True)
  File "/lore/mersoj2/cad-to-openmc-review/CAD_to_OpenMC/src/CAD_to_OpenMC/assembly.py", line 606, in stl2h5m
    self.check_h5m_file(h5m_file)
  File "/lore/mersoj2/cad-to-openmc-review/CAD_to_OpenMC/src/CAD_to_OpenMC/assembly.py", line 617, in check_h5m_file
    raise RuntimeError(
RuntimeError: Generated h5m-file does not appear to be a hdf-file
```

The data in the `output.h5m` indeed looks like a vtk file.

I have used b8d07bea6ac055d512d96ec5983f31a9bfce0f15 (v1.0.0).

https://github.com/openjournals/joss-reviews/issues/7710
