### 2025-02-20 openmsr/CAD_to_OpenMC#137 @nasserma
Also, whether I install the PyPI or development version using `pip`, a large number (from 3.6.1->4.6.1) of versions of `trimesh` are installed which is definitely not intended behavior, this warning from `pip` might provide some guidance,

> INFO: This is taking longer than usual. You might need to provide the dependency resolver with stricter constraints to reduce runtime. See https://pip.pypa.io/warnings/backtracking for guidance. If you want to abort this run, press Ctrl + C.
