### 2025-02-20 #7710 @nasserma
@ebknudsen Firstly, thanks for this open-source contribution, from reviewing the code it seems to have been a decent amount of work! However, there are some installation issues the need to be resolved before I can continue with the review,

https://github.com/openmsr/CAD_to_OpenMC/issues/137

but focusing on the manuscript and overall contribution I do have some initial comments/questions,

1. It is unclear from the manuscript and package documentation what is fundamentally being done by the package. My understanding is that a domain described using a standard CAD format (STEP) is meshed and then converted to an appropriate MCPT format using DAGMC. The summary, statement of need, and methods sections need to explicitly state what functionality that is missing from DAGMC is actually provided.
2. In the package documentation, two existing packages (`cad_to_dagmc` and `stellarmesher`) are stated to "solve the same problem as CAD_to_OpenMC, and in fact in many cases are interchangeable." Neither of these are referenced in the manuscript, which is very important, in that one of the requirements for publication in JOSS is,

_"significant contribution to the available open source software that either enables some new research challenges to be addressed or makes addressing research challenges significantly better (e.g., faster, easier, simpler)."_

How does this package meet either or both of these requirements if there are already two existing packages which address this issue?
3. Is there any particular reason (licensing, contribution issues, etc) that the functionality in this package could not be directly included in DAGMC?
4. Does the package provide automated functionality for users to understand the quality of mesh that is being provided to DAGMC? For example, is there functionality for the generation of the results in Figure 3 automatically? This would be important information for users to determine whether or not the conversion process was successful compared to a user-specified relative error (w.r.t. sub-domain volumes).
5. Regarding constraint A of the surface meshing process, is the absolute tolerance of the discretization automatically determined to meet this constraint?
6. Regarding constraint C of the surface meshing process, the description of how this case is being handled by the package is unclear.
