### 2025-02-20 openmsr/CAD_to_OpenMC#137 @ebknudsen
@nasserma Thanks for the review effort!
Let me first state that the issue with trimesh has cropped up after submitting to JOSS. I have little leverage on the trimesh-package, but we are working on the issue at present (see: https://github.com/united-neux/CAD_to_OpenMC/issues/3) Note that since submitting the development repo for CAD_to_OpenMC has moved. That is confusing but unfortunately was necessary due to company funding issues.
As a workaround, we will make the trimesh dependency optional, as it is not fundamental for the workings of CAD_to_OpenMC (The conda/mamba-installation route works fine however).
