### 2025-02-20 #7710 @ebknudsen
> [@ebknudsen](https://github.com/ebknudsen) Firstly, thanks for this open-source contribution, from reviewing the code it seems to have been a decent amount of work! However, there are some installation issues that need to be resolved before I can continue with the review,
>
> [openmsr/CAD_to_OpenMC#137](https://github.com/openmsr/CAD_to_OpenMC/issues/137)
>
> but focusing on the manuscript and overall contribution I do have some initial comments/questions,
>
>     1. It is unclear from the manuscript and package documentation what is fundamentally being done by the package. My understanding is that a domain described using a standard CAD format (STEP) is meshed and then converted to an appropriate MCPT format using DAGMC. The summary, statement of need, and methods sections need to explicitly state what functionality that is missing from DAGMC is actually provided.
>
>     2. In the package documentation, two existing packages (`cad_to_dagmc` and `stellarmesher`) are stated to "solve the same problem as CAD_to_OpenMC, and in fact in many cases are interchangeable." Neither of these are referenced in the manuscript, which is very important, in that one of the requirements for publication in JOSS is,
>
>
> _"significant contribution to the available open source software that either enables some new research challenges to be addressed or makes addressing research challenges significantly better (e.g., faster, easier, simpler)."_
>
> How does this package meet either or both of these requirements if there are already two existing packages which address this issue?
>
>     3. Is there any particular reason (licensing, contribution issues, etc) that the functionality in this package could not be directly included in DAGMC?
>
>     4. Does the package provide automated functionality for users to understand the quality of mesh that is being provided to DAGMC? For example, is there functionality for the generation of the results in Figure 3 automatically? This would be important information for users to determine whether or not the conversion process was successful compared to a user-specified relative error (w.r.t. sub-domain volumes).
>
>     5. Regarding constraint A of the surface meshing process, is the absolute tolerance of the discretization automatically determined to meet this constraint?
>
>     6. Regarding constraint C of the surface meshing process, the description of how this case is being handled by the package is unclear.

@nasserma Thanks for the thourough review! I will try to answer your questions below.

1. IMHO CAD_to_OpenMC does not add any functionality to DAGMC, it simply facilitates a convenient route to generating geometry that may be understood by the DAGMC-library. The DAGMC-library acts as an interface between a standardized triangular surface mesh format and a set of Monte Carlo particle transport codes. It is entirely possible to generate DAGMC-understandable geometry "by hand" albeit extremely inconvenient. Most users of DAGMC use a commercial product (cubit) - that is the situation we are trying to remedy. Most CAD-systems do not have the option of exporting geometry in the native DAGMC-format, but most support STEP.

2. The other packages were not referenced, since AFAIK they have not been published anywhere. CAD_to_OpenMC has features that the other packages do not have (yet). We will certainly add references to there code-repos if that can be considered appropriate. Stellarmesher is newer package than cad-to-dagmc and CAD_to_OpenMC, which came about almost at the same time. Neither of the other two have the feature-set that CAD_to_OpenMC has. This should be made more clear in the text - we will attempt to clarify what sets CAD_to_OpenMC apart.

3. Yes - I can think of 2. First, the process of generating a triangularized geometry can take a long time. In an extreme case (a model of the MSRE-reactor) this ran up to 120 minutes. Having to do this for every Monte Carlo-run, would be impractical. Second, If it is a separate entity it is easy to reuse geometry in a different location. E.g. the control rods in the GODIVA_IV-example. In this case the same triangularized geometry (same file) is included twice in different locations, with independent dynamic vertical translations, to create the 5 experimental cases.

4. DAGMC-specs allow for a single flag per file to be set wrt. mesh-tolerance, which is reported at runtime during an OpenMC-run. I do not believe this happens for MCNP however, but I don't know. CAD_to_OpenMC sets this flag. It is possible to set any number of extra flags in a DAGMC-file, but only this is reported runtime.
If available CAD_to_OpenMC calls trimesh to check for watertight mesh upon creation. If not the

5. No. CAD_to_OpenMC does not do this. To set this automatically would require to analyze the geometry and/or explicitly parse the STEP-geometry as it may be anything. This is simply beyond the scope of CAD_to_OpenMC. It would be conceivable to check if some triangles overlap between surfaces, which would give a user the info that the tolerance is not tight enough.

6. We acknowledge that this is unclear. We will expand a bit more on this process in the text.

I hope that clears things up.
cheers Erik
