### 2025-02-21 f223af8
fix datadirectory choice
otherwise an object could not be remeshed without crashing
