+++ src/CAD_to_OpenMC/assembly.py
@@ -496,10 +496,12 @@ class Assembly:
     def _datadir_name(self,h5m_filename=""):
         h5mf=pl.Path(h5m_filename)
         if self.datadir==".":
-            self.datadir=datetime.now().strftime(f"{h5mf.stem}_%Y%m%d_%H%M%S.%f")
+            datadir = datetime.now().strftime(f"{h5mf.stem}_%Y%m%d_%H%M%S.%f")
+        else:
+            datadir = self.datadir
         if (self.verbose):
-            print(f"INFO: storing temporary data in directory: {self.datadir}")
-        return self.datadir
+            print(f"INFO: storing temporary data in directory: {datadir}")
+        return datadir
 
     def solids_to_h5m(
         self,
