### 2025-02-21 58cab1d
1st stab at makign trimesh optional
