+++ requirements.txt
@@ -3,7 +3,6 @@ pyparsing
 cadquery>=2.2.0
 cadquery-ocp>=7.7.0
 numpy<2.0a0
-trimesh
 networkx
 unidecode
 Cython
