### 2025-02-24 d606490
update deps to numpy 2, which is now OK
