+++ requirements.txt
@@ -2,7 +2,7 @@ gmsh
 pyparsing
 cadquery>=2.2.0
 cadquery-ocp>=7.7.0
-numpy<2.0a0
+numpy
 networkx
 unidecode
 Cython
