### 2025-02-24 openmsr/CAD_to_OpenMC#127 @ebknudsen
@bonh @jacobmerson To provide a bit of context:
Indeed these other packages that try to solve the same problem, and yes some of us are contributing to more than one of them.
The history is that CAD_to_OpenMC and cad_to_dagmc are of similar age, stellarmesher somewhat newer. Partially stellarmesh was created with an ambition to unify efforts across projects, so far (due to funding and staffing issues) that has not been successful, and so the older packages retain value, and are being actively maintained.

That said - I agree with your comments and will clarify the differences (technical as well as stylistic) in in the text.
