### 2025-02-26 8c98e02
add a test for the sequential tag options as well
