### 2025-02-26 eeb433b
refactor tagging tests and harness
