### 2025-02-26 a873275
moving import_stp_files call to test run func
