### 2025-02-28 openmsr/CAD_to_OpenMC#136 @ebknudsen
@jacobmerson This happens if for some reason the hdf5-library is not available.
