### 2025-02-28 openmsr/CAD_to_OpenMC#131 @ebknudsen
Agreed - I will add something like that, in addition to the pytest-based automated tests, which are run for all PRs. Although I should mention that these are run at the "new" repo location united-neux/CAD_to_OpenMC
