### 2025-02-28 openmsr/CAD_to_OpenMC#137 @ebknudsen
This ought to now be fixed
