### 2025-02-28 openmsr/CAD_to_OpenMC#135 @ebknudsen
The root cause of this should now be fixed by #139.
