+++ bibliography.bib
@@ -162,13 +162,10 @@
         month = {jan}
 }
 
-@misc{icsbep_2020,
-	title = {{ICSBEP} Handbook 2020},
-	url = {https://www.oecd-ilibrary.org/nuclear-energy/data/international-criticality-safety-benchmark-evaluation-project-handbook/icsbep-handbook-2020_7e0ebc50-en},
-	doi = {10.1787/7e0ebc50-en},
-	abstract = {The evaluated criticality safety benchmark data in the 2020 edition contains 582 evaluations with benchmark specifications for 5 053 critical, near-critical or subcritical configurations, 45 criticality alarm placement/shielding configurations with multiple dose points for each, and 237 configurations which have been categorised as fundamental physics measurements that are relevant to criticality safety applications. New to the handbook are the first experiments from the Thermal/Epithermal {eXperiments} ({TEX}) program that were performed at the National Critical Experiments Research Center ({NCERC}) in the United States.},
-	publisher = {{OECD}},
-	urldate = {2024-07-03},
-	langid = {english},
-        date = {2020}
+@misc{icsbep_2022,
+        title = {International Handbook of Evaluated Criticality Safety Benchmark Experiments},
+        institution = {OECD Nuclear Energy Agency},
+        location = {Paris},
+        year = {2022/2023},
+        number = {7645}
 }
