### 2025-03-03 openmsr/CAD_to_OpenMC#134 @ebknudsen
- ICSBEP handbook link is broken, gives “withdrawn content error”

fixed by b398ded
