### 2025-03-03 openmsr/CAD_to_OpenMC#134 @ebknudsen
>
> - line 58, when you discuss surfaces, curves, etc. it may be useful to distinguish that OCCT is using a BREP rather than CSG.  > And describe how BREP makes use of topology and shape information. (this isn't required, but I don't think most neutronics > folks are familiar with how BREP and CSG models are different).
> - You are missing some aspects of the topology in your description, i.e., regions, shells, loops

I fully agree with your comment here - and have attempted to correct it in the text. This should be fixed by b2b5e359171d8c982d3a10453e9bc702516ec8af
