### 2025-03-03 openmsr/CAD_to_OpenMC#134 @ebknudsen
@jacobmerson I believe I have fixed all the issues in raised by this issue, except for step-files of all reactor models and CSG-volume.
