### 2025-03-03 7a898d9
add scripts+model for detailed godiva4
