+++ examples/godiva_iv/build_model_detailed.py
@@ -1,0 +1,454 @@
+import openmc
+import matplotlib.colors as mcolors
+from matplotlib import colormaps as cm
+import sys
+import math
+import pathlib as pl
+import numpy as np
+import itertools as it
+in2cm = 2.54
+
+fuels={
+"fuel_ring_101": ( ('U233',4.6624e-6),('U234',4.7311e-4),('U235',4.3069e-2),('U236', 3.1300e-4),('U238',2.3373e-3),('Mo',1.7013e-3) ),
+"fuel_ring_102": ( ('U233',4.6953e-6),('U234',4.7635e-4),('U235',4.3364e-2),('U236', 3.1521e-4),('U238',2.3631e-3),('Mo',1.7366e-3) ),
+"fuel_ring_103": ( ('U233',4.6795e-6),('U234',4.7439e-4),('U235',4.3186e-2),('U236', 3.1416e-4),('U238',2.3876e-3),('Mo',1.7209e-3) ),
+"fuel_ring_104": ( ('U233',4.6634e-6),('U234',4.7327e-4),('U235',4.3083e-2),('U236', 3.1308e-4),('U238',2.3332e-3),('Mo',1.6974e-3) ),
+"fuel_ring_105": ( ('U233',4.6427e-6),('U234',4.7116e-4),('U235',4.2892e-2),('U236', 3.1168e-4),('U238',2.3228e-3),('Mo',1.7104e-3) ),
+"fuel_ring_106": ( ('U233',4.6903e-6),('U234',4.7581e-4),('U235',4.3313e-2),('U236', 3.1488e-4),('U238',2.3652e-3),('Mo',1.7319e-3) ),
+"up_in_subassy_plate" : ( ('U233',4.7059e-6),('U234',4.7743e-4),('U235',4.3462e-2),('U236', 3.1593e-4),('U238',2.3684e-3),('Mo',1.7319e-3) ),
+"int_in_subassy_plate" : ( ('U233',4.5562e-6),('U234',4.6239e-4),('U235',4.2092e-2),('U236', 3.0587e-4),('U238',2.2795e-3),('Mo',1.7319e-3) ),
+"safety_block" : ( ('U233',4.6411e-6),('U234',4.7095e-4),('U235',4.2873e-2),('U236', 3.1158e-4),('U238',2.3266e-3),('Mo',1.7319e-3) ),
+"control_rod" : ( ('U233',4.6944e-6),('U234',4.7632e-4),('U235',4.3360e-2),('U236', 3.1516e-4),('U238',2.3580e-3),('Mo',1.7319e-3) ),
+"burst_rod" : ( ('U233',4.7807e-6),('U234',4.8507e-4),('U235',4.4157e-2),('U236', 3.2095e-4),('U238',2.4013e-3),('Mo',1.7319e-3) ),
+}
+
+steel303_composition=( ('C',3.0083e-4),('Si',1.7154e-3),('P',1.5554e-4,),('S',4.5067e-4),('Cr',1.6678e-2),('Mn',8.7693e-4),('Fe',6.0580e-2),('Ni',7.3878e-3),('Mo',1.5065e-4) )
+s303_bits={ nm:steel303_composition for nm in ["spindle","spindle_nut","safety_block_base" ,"clamp_support","long_align_pin", "belly_band", "belly_band_bolt", "locking_bolt", "steel_screen", "steel_invertedt", "steel"]}
+
+vascomax300_composition=(
[…]
