### 2025-03-03 27467ec
add scripts + steps 4 simplified
