+++ examples/godiva_iv/build_model_simplified.py
@@ -1,0 +1,471 @@
+from initialize_materials import create_materials
+import openmc
+import matplotlib.colors as mcolors
+from matplotlib import colormaps as cm
+import sys
+import math
+import pathlib as pl
+import numpy as np
+import itertools as it
+in2cm = 2.54
+
+def create_materials(operating_temp):
+    #fuel-like materials
+    safety_block = openmc.Material(name='safety_block', temperature=operating_temp)
+    safety_block.add_element('Mo',1.1399e-3,'ao')
+    safety_block.add_nuclide('U233',4.6384e-6,'ao')
+    safety_block.add_nuclide('U234',4.7068e-4,'ao')
+    safety_block.add_nuclide('U235',4.2848e-2,'ao')
+    safety_block.add_nuclide('U236',3.1140e-4,'ao')
+    safety_block.add_nuclide('U238',2.3253e-3,'ao')
+    safety_block.set_density('g/cc',7288/401.298)
+    safety_block.volume=401.298
+
+    ISP = openmc.Material(name='ISP', temperature=operating_temp)
+    ISP.add_element('Mo',1.7248e-3,'ao')
+    ISP.add_nuclide('U233',4.6326e-6,'ao')
+    ISP.add_nuclide('U234',4.7007e-4,'ao')
+    ISP.add_nuclide('U235',4.2791e-2,'ao')
+    ISP.add_nuclide('U236',3.1101e-4,'ao')
+    ISP.add_nuclide('U238',2.3249e-3,'ao')
+    ISP.set_density('g/cc',(3511.0+3312.0)/386.869)
+    ISP.volume=386.869
+    subassembly_plate_inner = ISP.clone()
+    subassembly_plate_inner.name = 'subassembly_plate_inner'
+
+    fuel_ring = openmc.Material(name='fuel_ring', temperature=operating_temp)
+    fuel_ring.add_element('Mo',1.7030e-3,'ao')
+    fuel_ring.add_nuclide('U233',4.6343e-6,'ao')
+    fuel_ring.add_nuclide('U234',4.7016e-4,'ao')
+    fuel_ring.add_nuclide('U235',4.2801e-2,'ao')
+    fuel_ring.add_nuclide('U236',3.1112e-4,'ao')
+    fuel_ring.add_nuclide('U238',2.3328e-3,'ao')
+    fuel_ring_mass=8378.0+8134.0+8046.0+6909.0+8166.0+9283.0
+    fuel_ring.set_density('g/cc',fuel_ring_mass/2782.931)
+    fuel_ring.volume=2782.931
+
+    control_rod = openmc.Material(name='control_rod', temperature=operating_temp)
+    control_rod.add_element('Mo',1.5550e-3,'ao')
+    control_rod.add_nuclide('U233',4.7315e-6,'ao')
+    control_rod.add_nuclide('U234',4.8008e-4,'ao')
+    control_rod.add_nuclide('U235',4.3703e-2,'ao')
+    control_rod.add_nuclide('U236',3.1765e-4,'ao')
+    control_rod.add_nuclide('U238',2.3767e-3,'ao')
+    control_rod.set_density('g/cc',824.0/44.88)
+    control_rod.volume=44.88
+
+    burst_rod = openmc.Material(name='burst_rod', temperature=operating_temp)
+    burst_rod.add_element('Mo',1.5909e-3,'ao')
+    burst_rod.add_nuclide('U233',4.8405e-6,'ao')
+    burst_rod.add_nuclide('U234',4.9114e-4,'ao')
+    burst_rod.add_nuclide('U235',4.4710e-2,'ao')
+    burst_rod.add_nuclide('U236',3.2497e-4,'ao')
[…]
