### 2025-03-03 dd61212
remove unused type flag
