### 2025-03-03 openmsr/CAD_to_OpenMC#133 @ebknudsen
7a898d967775008caadb6949664d3d03adfbbbc and 27467ecd123878ef297bca0854018d4159e5 include stepfiles and scripts for tables 1 and 2.
