### 2025-03-03 openmsr/CAD_to_OpenMC#133 @ebknudsen
The scripts also include provisions for volume calculations
