### 2025-03-03 469e7df
add reuqirements to README
