### 2025-03-03 openmsr/CAD_to_OpenMC#128 @ebknudsen
Consider it done.

fixed by 469e7dfe4496fb544947b58e199c6dcd59fc23c
