### 2025-03-03 7b761fe
add comment about usage of MC-neutronics
