+++ paper.md
@@ -41,8 +41,9 @@ DAGMC-library, which is usable by the main particle-transport packages.
 
 # Statement of need
 Most Monte Carlo neutronics-, and general particle transport (e.g. OpenMC, GEANT4, MCNP, Fluka [@openmc_2013;
-@geant4_2003; @mcnp_2022; @fluka_2005; @fluka_2014]) software packages use the
-concept of Constructive Solid Geometry (CSG) to the describe the geometry or
+@geant4_2003; @mcnp_2022; @fluka_2005; @fluka_2014]) software packages are ubiquitous in nuclear science. They're used to
+assess performance of the system, estimate radiation levels, shielding requirements, among many other uses.
+Almost all of them use the concept of Constructive Solid Geometry (CSG) to the describe the geometry or
 scene in which a computations are to performed. Computations generally amount
 to tracing a large number of rays through the geometry. Such models range in
 complexity from simple scenes of a only a few geometrical objects to several
