### 2025-03-03 openmsr/CAD_to_OpenMC#132 @ebknudsen
Good idea!
7b761fe  fixes this
