### 2025-03-03 9cf1b85
add rudimentary policy about contributing
