### 2025-03-03 openmsr/CAD_to_OpenMC#130 @ebknudsen
9cf1b855947c1eee2acfb90b34352ddf58c711f8 adds a rudimentary statement about this to the README. Thus far, contributions and report have come in as issues/PRs by themselves, so it has not been necessary.
