### 2025-03-03 openmsr/CAD_to_OpenMC#136 @ebknudsen
That is a very good idea. I will attempt to add such an assert ASAP
