### 2025-03-03 5e4d484
fix example script
