### 2025-03-04 a25ce77
fix recursion error
move hdf5_in_moab outside Assembly class
and let the class use the underlying fuction instead.
This to avoid infinite recursion
