### 2025-03-04 openmsr/CAD_to_OpenMC#136 @ebknudsen
This was fixed by #140
