### 2025-03-04 49022d5
add a self-contained test
