+++ tests/test_self_contained.py
@@ -1,0 +1,250 @@
+import CAD_to_OpenMC.assembly as ab
+import subprocess as sp
+import h5py
+import numpy as np
+import pathlib as pl
+
+instep = """
+ISO-10303-21;
+HEADER;
+FILE_DESCRIPTION(('FreeCAD Model'),'2;1');
+FILE_NAME('Open CASCADE Shape Model','2025-03-04T11:39:47',('Author'),(
+    ''),'Open CASCADE STEP processor 7.7','FreeCAD','Unknown');
+FILE_SCHEMA((
+'AP242_MANAGED_MODEL_BASED_3D_ENGINEERING_MIM_LF. {1 0 10303 442 1 1 4 
+}'));
+ENDSEC;
+DATA;
+#1 = APPLICATION_PROTOCOL_DEFINITION('international standard',
+  'ap242_managed_model_based_3d_engineering',2013,#2);
+#2 = APPLICATION_CONTEXT('Managed model based 3d engineering');
+#3 = SHAPE_DEFINITION_REPRESENTATION(#4,#10);
+#4 = PRODUCT_DEFINITION_SHAPE('','',#5);
+#5 = PRODUCT_DEFINITION('design','',#6,#9);
+#6 = PRODUCT_DEFINITION_FORMATION('','',#7);
+#7 = PRODUCT('cuboid','cuboid','',(#8));
+#8 = PRODUCT_CONTEXT('',#2,'mechanical');
+#9 = PRODUCT_DEFINITION_CONTEXT('part definition',#2,'design');
+#10 = ADVANCED_BREP_SHAPE_REPRESENTATION('',(#11,#15),#165);
+#11 = AXIS2_PLACEMENT_3D('',#12,#13,#14);
+#12 = CARTESIAN_POINT('',(0.,0.,0.));
+#13 = DIRECTION('',(0.,0.,1.));
+#14 = DIRECTION('',(1.,0.,-0.));
+#15 = MANIFOLD_SOLID_BREP('',#16);
+#16 = CLOSED_SHELL('',(#17,#57,#97,#119,#141,#153));
+#17 = ADVANCED_FACE('',(#18),#52,.F.);
+#18 = FACE_BOUND('',#19,.F.);
+#19 = EDGE_LOOP('',(#20,#30,#38,#46));
+#20 = ORIENTED_EDGE('',*,*,#21,.F.);
+#21 = EDGE_CURVE('',#22,#24,#26,.T.);
+#22 = VERTEX_POINT('',#23);
+#23 = CARTESIAN_POINT('',(0.,0.,0.));
+#24 = VERTEX_POINT('',#25);
+#25 = CARTESIAN_POINT('',(0.,0.,16.));
+#26 = LINE('',#27,#28);
+#27 = CARTESIAN_POINT('',(0.,0.,0.));
+#28 = VECTOR('',#29,1.);
+#29 = DIRECTION('',(0.,0.,1.));
+#30 = ORIENTED_EDGE('',*,*,#31,.T.);
+#31 = EDGE_CURVE('',#22,#32,#34,.T.);
+#32 = VERTEX_POINT('',#33);
+#33 = CARTESIAN_POINT('',(0.,8.,0.));
+#34 = LINE('',#35,#36);
+#35 = CARTESIAN_POINT('',(0.,0.,0.));
+#36 = VECTOR('',#37,1.);
+#37 = DIRECTION('',(-0.,1.,0.));
+#38 = ORIENTED_EDGE('',*,*,#39,.T.);
+#39 = EDGE_CURVE('',#32,#40,#42,.T.);
[…]
