### 2025-03-04 985bf11
add doc for self conatined test
