### 2025-03-04 openmsr/CAD_to_OpenMC#131 @ebknudsen
There is now a self-contained test script in the tests directory with exactly this purpose, which checks that an actual geometry, albeit a very simple one, get triangularized in the expected way.
