+++ docs/api.rst
@@ -1,0 +1,7 @@
+API
+===
+
+../src autosummary::
+   :toctree: generated
+
+   CAD_to_OpenMC
