### 2025-03-04 2c96317
update docs - mostly copy-paste from README
