+++ docs/source/api.rst
@@ -5,3 +5,4 @@ API
    :toctree: generated
 
    CAD_to_OpenMC
+   CAD_to_OpenMC.assembly
