+++ docs/source/examples.rst
@@ -1,6 +1,0 @@
-Examples
-====================
-
-We supply a few example cases that can be meshed using CAD_to_OpenMC
-1. A 7-pin cell where eac pin consists of a cylindrical rod with a cylindrical hollow inside, surrounded by a cladding cylinder - slightly shorter that the core rod.
-2. A full-scale model of the MSRE 
