### 2025-03-04 45416cb
yet another update to reqs to build the docs
