+++ docs/api.rst
@@ -1,7 +1,0 @@
-API
-===
-
-../src autosummary::
-   :toctree: generated
-
-   CAD_to_OpenMC
