+++ .readthedocs.yaml
@@ -23,9 +23,13 @@ sphinx:
 # formats:
 #    - pdf
 
+
+conda:
+   file: docs/environment.yml
+
 # Optionally declare the Python requirements required to build your docs
-python:
-   install:
-   - requirements: docs/requirements.txt
-   - method: pip
-     path: .
+##python:
+#   install:
+#   - requirements: docs/requirements.txt
+#   - method: pip
+#     path: .
