+++ .readthedocs.yaml
@@ -25,8 +25,8 @@ sphinx:
 
 
 conda:
-   environment: test2
    file: docs/environment.yml
+      environment: test2
 
 # Optionally declare the Python requirements required to build your docs
 ##python:
