+++ docs/environment.yml
@@ -1,7 +1,7 @@
 name: test2
 channels:
-- conda-forge
+  - conda-forge
 dependencies:
-- cad-to-openmc
-- python=3.10
+  - cad-to-openmc
+  - python=3.10
 
