+++ .readthedocs.yaml
@@ -25,7 +25,7 @@ sphinx:
 
 
 conda:
-   file: docs/environment.yml
+   environment: docs/environment.yml
 
 # Optionally declare the Python requirements required to build your docs
 ##python:
