+++ .readthedocs.yaml
@@ -9,7 +9,7 @@ version: 2
 build:
    os: ubuntu-22.04
    tools:
-     python: "3.10"
+     python: "mambaforge-latest"
      # You can also specify other tool versions:
      # nodejs: "16"
      # rust: "1.55"
