+++ docs/source/index.rst
@@ -1,5 +1,5 @@
 Welcome to The Documentation for CAD_to_OpenMC
-=============================================
+==============================================
 **CAD_to_OpenMC** is a python tool intened to ease the process of
 creating OpenMC-ready surface meshes from a CAD-generated representation in the form
 of a .step (or .stp)-file.
