+++ docs/conf.py
@@ -6,7 +6,7 @@ project = "CAD_to_OpenMC Project"
 copyright = "2025, CAD_to_OpenMC-team"
 author = "ebknudsen the Docs core team"
 
-master_doc = 'index'
+master_doc = 'source/index.rst'
 
 extensions = [
     "sphinx.ext.duration",
