+++ docs/index.rst
@@ -1,11 +1,11 @@
-Welcome to The Documentation for CAD_to_OpenMC
+Welcome to The documentation for CAD_to_OpenMC
 ==============================================
 **CAD_to_OpenMC** is a python tool intened to ease the process of
 creating OpenMC-ready surface meshes from a CAD-generated representation in the form
 of a .step (or .stp)-file.
 
-Check out the :doc:`usage` section for further information, including
-:doc:`install`.
+Check out the :doc:`source/usage` section for further information, including
+:doc:`source/install`.
 
 .. note::
 
@@ -16,7 +16,7 @@ Contents
 
 .. toctree::
 
-   install.md
-   usage.md
-   api.rst
+   source/install.md
+   source/usage.md
+   source/api.rst
 
