+++ docs/index.rst
@@ -1,15 +1,15 @@
 Welcome to The documentation for CAD_to_OpenMC
 ==============================================
-**CAD_to_OpenMC** is a python tool intened to ease the process of
-creating OpenMC-ready surface meshes from a CAD-generated representation in the form
-of a .step (or .stp)-file.
+**CAD_to_OpenMC** is a python tool, made to ease the process of
+creating OpenMC-ready surface meshes from a CAD-generated representationof the geometry
+in the form of a .step (or .stp)-file.
 
 Check out the :doc:`source/usage` section for further information, including
 :doc:`source/install`.
 
 .. note::
 
-   This project is under active development.
+   This project (and its documentation)is under active development.
 
 Contents
 --------
