### 2025-03-06 openmsr/CAD_to_OpenMC#129 @ebknudsen
@bonh The docs that were there were an artifact of an early attempt at using readthedocs, which was not finished. They have been updated have may be browsed at https://cad-to-openmc.readthedocs.io/
