### 2025-03-06 openmsr/CAD_to_OpenMC#129 @ebknudsen
@bonh Thanks for commenting on that. That way it got fixed. Closing.
