### 2025-03-06 openmsr/CAD_to_OpenMC#130 @ebknudsen
closing as fixed
