### 2025-03-06 db53a05
adress #127 and put c2o in relief to others
