+++ bibliography.bib
@@ -169,3 +169,15 @@
         year = {2022/2023},
         number = {7645}
 }
+
+@misc{cadtodagmc,
+        author = {Jon Shimwell},
+        url = {https://github.com/fusion-energy/cad_to_dagmc},
+        year = {2025}
+}
+
+@software{stellarmesh,
+    author = {Koen, A., et.al.},
+    url = {https://github.com/stellarmesh/stellarmesh},
+    year = {2025}
+}
