### 2025-03-06 openmsr/CAD_to_OpenMC#127 @ebknudsen
@bonh @jcaobmerson I have now tried to put CAD_to_OpenMC and it's design philosophy, in relief to the others. I do hope that is what you were after.
