### 2025-03-06 325529d
add a little more deail about other packages
