### 2025-03-06 openmsr/CAD_to_OpenMC#134 @ebknudsen
Step-files of all reactor models is covered by #133.

Upon considering the idea of using CSG-models to also compute the volumes of all the parts - I don't think this is very useful  as a general exercise, unless tolerances for the computation is extremely tight, meaning very long runs of OpenMC.
It makes more sense to do this for a few chosen objects. I've created #144 to track this.
