### 2025-03-06 openmsr/CAD_to_OpenMC#134 @ebknudsen
Thus this issue ought to be done - closing soon.
