### 2025-03-07 #7710 @bonh
@ebknudsen I suggest that you leave issues concerning JOSS stuff open until we are done. I keep loosing track... 😊
