### 2025-03-10 openmsr/CAD_to_OpenMC#133 @ebknudsen
143877a updates the text to include references to scripts and datasets for reproducing table 3. Datasets are published on Zenodo - they tend to be bigger than what is comfortably handled by git.
