### 2025-03-11 openmsr/CAD_to_OpenMC#128 @bonh
> Starting from vanilla conda env required around 340 MB of download. Just want to make sure that this is on purpose: Are all these packages pulled by conda really required?

Can you comment on that one?
