### 2025-03-11 openmsr/CAD_to_OpenMC#131 @bonh
When I try to execute the test script I get `ModuleNotFoundError: No module named 'h5py'`. I installed the code via conda so not from branch `paper_joss` but I guess that `h5py` is required for other stuff too and then it should be part of the conda recipe, right?
