### 2025-03-11 openmsr/CAD_to_OpenMC#130 @bonh
Would you like to add a sentence whether users are allowed to ask questions, perhaps as issues?
