### 2025-03-11 openmsr/CAD_to_OpenMC#127 @bonh
I like the clarification. Few questions/suggestions:

> First, it is aimed at working for all step-geometries, with no assumptions on geometry.

I don't get what you mean by "all step-geometries" and "no assumptions on geometry".

> Second, it must be relatively easy to add code for a new backend should one wish to do so. From this requirement stems the code-structure as it is, with one frontend and several backend-classes. This may be seen as a measure to enhance the longevity of the project.

Do you mean by "backend" the meshing backends listed in the readme? Perhaps you'd like to clarify this. Further, if you think this is important for your package, perhaps you'd like to change the project structure to more strongly reflect the "frontend"-"backend" relationship? And finally, if you aim for "easy to add code for a new backend" I suggest that you provide a specific example for the user how to do so (subclassing `assemblymesher`, filling in the required methods, and so on).

> Last, it must be easy to extract, generate, and manipulate material tags from the underlying step-model. As models become large and complex, maintaining a separate list of materials, in sequence, is seen as cumbersome and error prone.

Just playing devils advocate here (and I haven't really tried out the other software packages): The other software packages are not able to "extract, generate, and manipulate material tags from the underlying step-model"? To me that sounds like a fundamental functionality of a software package claiming to transform step files into something else.
