### 2025-03-11 #7710 @bonh
@ebknudsen I think the quality of writing can be improved and the list of references doesn't look perfect yet.
