### 2025-03-12 openmsr/CAD_to_OpenMC#131 @ebknudsen
You are entirely right - I missed this one since other stuff depends on it. This has been taken care of by #145
