### 2025-03-12 openmsr/CAD_to_OpenMC#130 @ebknudsen
Good point! 112b90cd31e7ddc6e0c5fcb5040a7ae5c96085d adds this.
