### 2025-03-17 openmsr/CAD_to_OpenMC#128 @bonh
What about `moab`?
