### 2025-03-17 openmsr/CAD_to_OpenMC#131 @bonh
It seems that installing `CadQuery` via pip misses some depedencies? I installed `CAD_to_OpenMC` from source using branch `joss_paper`. Running `test_self_contained.py` results in `ModuleNotFoundError: No module named 'typing_extensions'` which is used by `CadQuery` as far as I understand. Installing `typing_extensions` resolves this issue but then `ModuleNotFoundError: No module named 'multimethod'` . I now that this is more a `CadQuery` issue but it seems to me, that the `CadQuery` people recommend Conda for installing their software?
