### 2025-04-02 openmsr/CAD_to_OpenMC#136 @ebknudsen
Can we (re) close this?
