### 2025-04-02 openmsr/CAD_to_OpenMC#131 @ebknudsen
This was indeed an upstream issue, related to numpy transitioning from 1.x to 2.y which was not entirely reflected in cadquery's pip-stream. It works now with the new release (pip: 1.0.1-2), btw.
