### 2025-04-02 openmsr/CAD_to_OpenMC#129 @ebknudsen
Yes, for sure, this is work in progress. There's now an open issue on that subject: https://github.com/united-neux/CAD_to_OpenMC/issues/21
