### 2025-04-02 openmsr/CAD_to_OpenMC#128 @ebknudsen
moab is indeed necessary - at least for now. For the future some colleagues and I are working on a replacement base solely on h5py. It amount to reverse-engineering the moab-infrastructure with next to no documentation however, so it is slow going.
