### 2025-04-02 openmsr/CAD_to_OpenMC#128 @ebknudsen
For sure the biggest dependency will likely be OpenCASCADE, which is used by cadquery (and gmsh).
Of course numpy is a fairly large on at that, but is pretty likely to already be installed/downloaded anyway (IMHO).

The inclusion of gmsh creates a dep. on some Xorg libs, which are not at all needed by this project. One would have to repackage gmsh to not include its GUI do avoid that, but I would consider that out of scope for this project.
