### 2025-04-03 openmsr/CAD_to_OpenMC#127 @bonh
That was very helpful, thanks!
