### 2025-04-03 openmsr/CAD_to_OpenMC#128 @bonh
Perhaps you'd like to add moab to the list of dependencies for now. But you are mentioning it in the readme (https://github.com/openmsr/CAD_to_OpenMC/tree/paper_joss#to-installset-up-in-a-virtual-python-environment). Your choice.

Thanks!
