### 2025-04-03 #7710 @bonh
@mbarzegary to what level of detail is an API documentation necessary for publication in JOSS?

I'm trying to answer this question: "Functionality documentation: Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?"

See https://github.com/openmsr/CAD_to_OpenMC/issues/129
