### 2025-04-03 openmsr/CAD_to_OpenMC#129 @bonh
Not completely sure about that one but JOSS requires some "satisfactory level" of documentation and gives "API method documentation" as an example. I like to see the "full" API online but might not be necessary for JOSS. I'll ask our editor here: https://github.com/openjournals/joss-reviews/issues/7710#issuecomment-2774753879
