### 2025-04-03 openmsr/CAD_to_OpenMC#131 @bonh
What do you mean by "(pip: 1.0.1-2)"? Isn't that version like 15 years ago (https://pypi.org/project/pip/#history) ? I'm using version 25 something.
