### 2025-04-03 openmsr/CAD_to_OpenMC#131 @bonh
I have trouble installing cadquery properly; it works with cadquery=2.2.0 and python=3.10, but not with the newest version of cadquery. What version of cadquery are you using?
