### 2025-04-03 #7710 @mbarzegary
@bonh I would say it depends on the type of the submission. If it's a library/module, then at least a list of available APIs is needed, but in case of a tool featuring a CLI/GUI, the API documentation would not be of a great importance.

Since this submission is a Python module and the main usage is through the APIs, I would say that the authors should at least provide the list of APIs and can complete the documentation of each function/class gradually later.
