### 2025-04-03 openmsr/CAD_to_OpenMC#129 @ebknudsen
I saw the clarifying note by the editor (thanks for asking btw.), and I acknowledge that documenting to the level required by JOSS requires some work on my part. This will take a day or two I expect.
That said - it is the intention to fully "read-the-docs"-ify the API-docs to cover all functions/methods. I'd like to see that myself.
