### 2025-04-06 openmsr/CAD_to_OpenMC#146 @jacobmerson
JOSS: minor paper typos
1. line 43 and 44 missing space before references
2. I think the following statement on line 125 is missing a word "but for with CAD"
3. line 192 "We submit that have shown that the tool presented...", missing a word
