### 2025-04-06 #7710 @jacobmerson
@ebknudsen a few minor typos mentioned in https://github.com/openmsr/CAD_to_OpenMC/issues/146 and I think it would be good for you to mention that conda forge may be needed in the installation docs as this would have saved me a lot of time.

And, I'd like to confirm that when the review is finalized you are planning to merge the `paper_joss` branch into main.

Otherwise, you have sufficiently addressed my concerns.
