### 2025-04-07 openmsr/CAD_to_OpenMC#131 @bonh
Installing it via https://github.com/openmsr/CAD_to_OpenMC/issues/125#issuecomment-2781226820 works! 👍

Then added `pip install . ` to actually install the `paper_joss` branch on top.
