### 2025-04-07 openmsr/CAD_to_OpenMC#131 @bonh
Now I run `python3 test_self_contained.py` and the output is
```
INFO: cuboid001_tmp.step imported - scaling
Info    :  - Label 'Shapes/cuboid' (3D)
Info    :  - Color (0.8, 0.8, 0.8) (3D & Surfaces)
INFO: reordering volumes after merge
INFO: storing temporary data in directory: cuboid001_tmp_20250407_134850.748766
INFO: triangulating solid 0
SUMMARY: solid_id material_tag     stl-file
SUMMARY:        1 cuboid           vol_1_face0000.stl vol_1_face0001.stl vol_1_face0002.stl vol_1_face0003.stl vol_1_face0004.stl vol_1_face0005.stl
INFO: reassembling stl-files into h5m structure
INFO: writing geometry to h5m: "cuboid001_tmp.h5m".
INFO: writing geometry to vtk: cuboid001_tmp.vtk
```
I guess it is working as expected? If so, I propose that you add a more meaningful message at the end.
