### 2025-04-07 #7710 @ebknudsen
> [@ebknudsen](https://github.com/ebknudsen) a few minor typos mentioned in [openmsr/CAD_to_OpenMC#146](https://github.com/openmsr/CAD_to_OpenMC/issues/146) and I think it would be good for you to mention that conda forge may be needed in the installation docs as this would have saved me a lot of time.
>
> And, I'd like to confirm that when the review is finalized you are planning to merge the `paper_joss` branch into main.
>
> Otherwise, you have sufficiently addressed my concerns.

@jacobmerson Thank you again for taking the time to review this contribution. I will incorporate a mention of conda-forge. I am really sorry that you had to spend extra time on account of this.
I can confirm that edits will be merged into the main branch.
