### 2025-04-07 openmsr/CAD_to_OpenMC#147 @bonh
JOSS: Examples
The paper describes two awesome looking examples: GODIVA IV and Molten Salt Reactors. I found a few files named accordingly and a folder `CAD_to_OpenMC/examples/godiva_iv`. I tried to run a few files but that doesn't seem to work properly out-of-the-box.  For example, to run `create_mesh_sample.py` I have to copy `7pin.step`from `tests`. Then I tried files in the folder `godiva_iv` but there was stuff missing.

In the documentation (https://cad-to-openmc.readthedocs.io/usage.html#advanced-example) you are using a different code to run an example. Is that code part of this submission?

I think it is necessary to have a few (a simple and a complex) example case documented (like in the paper, only online as part of the software documentation) and readily executable with a description of expected results (like in the paper).

https://github.com/openjournals/joss-reviews/issues/7710
