### 2025-04-07 openmsr/CAD_to_OpenMC#131 @ebknudsen
Thanks for the suggestion.  I've added a line that adds a message to that effect in the end.
