### 2025-04-07 openmsr/CAD_to_OpenMC#147 @ebknudsen
> The paper describes two awesome looking examples: GODIVA IV and Molten Salt Reactors. I found a few files named accordingly and a folder `CAD_to_OpenMC/examples/godiva_iv`. I tried to run a few files but that doesn't seem to work properly out-of-the-box. For example, to run `create_mesh_sample.py` I have to copy `7pin.step`from `tests`. Then I tried files in the folder `godiva_iv` but there was stuff missing.

The intention of the godiva-iv example is definitely for it to be complete. Did you use the zenodo-version?

>
> In the documentation (https://cad-to-openmc.readthedocs.io/usage.html#advanced-example) you are using a different code to run an example. Is that code part of this submission?

Not really. It is a simple script that runs CAD_to_OpenMC if necessary, and then runs various forms of OpenMC-calculations using that geometry. That said, I agree that in the very least, the essence of the script should be quoted in the online example to be more illustrative. I will try to fix that.

>
> I think it is necessary to have a few (a simple and a complex) example case documented (like in the paper, only online as part of the software documentation) and readily executable with a description of expected results (like in the paper).
>
> [openjournals/joss-reviews#7710](https://github.com/openjournals/joss-reviews/issues/7710)

I agree - I'll make sure that is available.
