### 2025-04-14 openmsr/CAD_to_OpenMC#147 @bonh
Mhh, no, I used work from the `examples` folder. Is there a reference to a zonodo repo somewhere (and I missed it, sry)?
