### 2025-04-14 #7710 @bonh
I'm almost done. Only waiting for a nice example, see https://github.com/openmsr/CAD_to_OpenMC/issues/147, to test the functionality and performance.
