### 2025-04-14 openmsr/CAD_to_OpenMC#147 @bonh
That would be awesome!
