### 2025-04-23 #7710 @ebknudsen
> I'm almost done. Only waiting for a nice example, see [openmsr/CAD_to_OpenMC#147](https://github.com/openmsr/CAD_to_OpenMC/issues/147), to test the functionality and performance.

I'm now done with including the two examples in the online docs, so all should now work.
