### 2025-04-23 openmsr/CAD_to_OpenMC#147 @bonh
How can I install the backend `db`?  In the docs it says that `it can be tricky`. So I tried `gmsh` and get:
```
Traceback (most recent call last):
  File "/tmp/CAD_to_OpenMC/examples/godiva_iv/step_conversion.py", line 17, in <module>
    B.run(backend='gmsh', h5m_filename='GIV_CR_det.h5m')
  File "/home/cb51neqa/miniconda3/envs/openmc-env/lib/python3.9/site-packages/CAD_to_OpenMC/assembly.py", line 241, in run
    self.solids_to_h5m(backend=backend, h5m_filename=h5m_filename, **kwargs)
  File "/home/cb51neqa/miniconda3/envs/openmc-env/lib/python3.9/site-packages/CAD_to_OpenMC/assembly.py", line 511, in solids_to_h5m
    stl_list = meshgen.generate_stls()
  File "/home/cb51neqa/miniconda3/envs/openmc-env/lib/python3.9/site-packages/CAD_to_OpenMC/assemblymesher_gmsh.py", line 153, in generate_stls
    dim,vid=vols[k]
IndexError: list index out of range
```
