### 2025-04-23 openmsr/CAD_to_OpenMC#147 @bonh
Running "scripts/step_to_h5m_zpre.py" doesn't seem to produce any files.

edit: just found an `exit()` in the script. Now it's running...

edit2: still running on my laptop 😄

edit3: It's using a lot of memory, greater 20 gb. Perhaps you'd like to put a warning in the docs?
