### 2025-04-23 openmsr/CAD_to_OpenMC#147 @ebknudsen
Me bad - I missed that  - the exit()-bug is now been fixed.

In addition, I've added a warning like you suggested. That reactor model *is* rather complicated, and has some thin shells with large radii, meaning the mesh tolerances have to be set low. This leads to a big memory footprint, and longish runtimes.
