### 2025-04-24 openmsr/CAD_to_OpenMC#147 @bonh
> Me bad - I missed that - the exit()-bug is now been fixed.
>
> In addition, I've added a warning like you suggested. That reactor model _is_ rather complicated, and has some thin shells with large radii, meaning the mesh tolerances have to be set low. This leads to a big memory footprint, and longish runtimes.

Do you have a rough idea how long it should take?

Edit: Runs for almost 3 h now; I'm not complaining just wanted to let you know. As soon as that test case is done, the results look good and we can make the other one run I'm done with the review
