### 2025-04-24 openmsr/CAD_to_OpenMC#147 @ebknudsen
That kind of runtime unfortunately to be expected for a geometry of this kind with "gmsh". I just ran it yesterday on my desktop, but didn't really time it. I guess a warning about this as well is warranted.

Great to hear about the review. I must say , this has been the most pleasant form of review I've ever encountered. It just makes everything better.
