### 2025-04-24 openmsr/CAD_to_OpenMC#147 @bonh
6 h now 😅

`INFO: GMSH generate surface mesh`
