### 2025-04-24 openmsr/CAD_to_OpenMC#147 @ebknudsen
yeah - sorry about that. This I cannot do anything about I'm afraid. I restarted as well with a time-command on it and I am at ```GMSH generate``` as well.
It will get there though.
