### 2025-04-25 openmsr/CAD_to_OpenMC#147 @bonh
Did it finish? How long did it take? Towards the end, I ran out of hard drive space 😒
I'll try again this evening and will let it run over night. Hopefully, I can have a nice view of the resulting mesh on Saturday.

I'm not able to run the first example. Please have a look at the error I posted above.
