### 2025-04-25 openmsr/CAD_to_OpenMC#147 @ebknudsen
Ok - thanks for checking - just checked my run - it clocked in at 11h (sic). I suspect a somewhat smaller example would have been better. It is much faster using another meshing backend than gmsh though.

Wrt. to the other issue - I will deal with that today. Don't know what went wrong now, as it did work for me. Please note that I do appreciate all the work you've put into this.
