### 2025-04-25 openmsr/CAD_to_OpenMC#147 @ebknudsen
> > How can I install the backend `db`? In the docs it says that `it can be tricky`.

The "tricky" comment refers to an issue about staggered versions of cadquery, cadqeury-ocp, OCCT, and numpy between the conda-world and the pip-world. Cadquery seems to prefer pip whereas the others prefer conda. The 'db' backend is q quite recent addition to OCCT which was incompatible with cadquery - in the pip-universe. In the conda-world something else clashed. (To add to the confusion cadquery-ocp is called merely OCP in conda).
All that said - gmsh-ing this one should of course work. I'll look into the matter now.

> > So I tried `gmsh` and get:
> > ```
> > Traceback (most recent call last):
> >   File "/tmp/CAD_to_OpenMC/examples/godiva_iv/step_conversion.py", line 17, in <module>
> >     B.run(backend='gmsh', h5m_filename='GIV_CR_det.h5m')
> >   File "/home/cb51neqa/miniconda3/envs/openmc-env/lib/python3.9/site-packages/CAD_to_OpenMC/assembly.py", line 241, in run
> >     self.solids_to_h5m(backend=backend, h5m_filename=h5m_filename, **kwargs)
> >   File "/home/cb51neqa/miniconda3/envs/openmc-env/lib/python3.9/site-packages/CAD_to_OpenMC/assembly.py", line 511, in solids_to_h5m
> >     stl_list = meshgen.generate_stls()
> >   File "/home/cb51neqa/miniconda3/envs/openmc-env/lib/python3.9/site-packages/CAD_to_OpenMC/assemblymesher_gmsh.py", line 153, in generate_stls
> >     dim,vid=vols[k]
> > IndexError: list index out of range
> > ```
>
> And what about that error? Is that something on my end?

I _think_ that this is caused by gmsh failing to create a mesh of some surface, but I'll have to look into the matter.
