### 2025-04-28 openmsr/CAD_to_OpenMC#147 @bonh
Did a new setup and it worked fine 😄
