### 2025-04-28 openmsr/CAD_to_OpenMC#147 @ebknudsen
Awesome! Great news. Thanks for the patience!
