### 2025-04-29 #7710 @ebknudsen
> [@mbarzegary](https://github.com/mbarzegary) [@ebknudsen](https://github.com/ebknudsen) I added issues regarding a few very minor typos and omissions in the current documentation, otherwise my review is complete.

These things have now been handled by [20bdd90](https://github.com/openmsr/CAD_to_OpenMC/commit/20bdd903ef20a6d19a3a9d47f54dc249a338758e) and [a9e8008](https://github.com/openmsr/CAD_to_OpenMC/commit/a9e80084629e2f9c747b56efeec1f4e13264ba03).
