### 2025-05-06 #7710 @ebknudsen
remaining refs with missing DOI are references to other github-hosted software projects which do not have a relevant title nor DOI
