### 2025-05-06 #7710 @mbarzegary
@ebknudsen given the green light of the reviewers, we will now work towards processing this for acceptance in JOSS. So please
* Check and merge my PR with bunch of editorial edits
* Work on the author's points of the final checklist I created ☝️
