### 2025-05-07 #7710 @ebknudsen
DOI for release v1.0.0-1 on Zenodo: https://doi.org/10.5281/zenodo.15358500

Authors, ORCID (where applicable), and License match those in the code and paper.
