### 2025-05-15 #7710 @ebknudsen
@mbarzegary I have now changed the zenodo archive to reflect version 1.0.0-1 and the paper title. Looks good to me!
