### 2025-05-21 #7710 @kyleniemeyer
@ebknudsen apologies for only raising this issue now, but your paper is too long—we suggest a target of 1000 words for JOSS papers, and 2000 words is an absolute upper limit. (This is because the software is the main focus, and JOSS does not publish research results. Extensive examples and benchmarking should go with documentation, rather than the paper.)

Can you please trim your paper down below that limit and ideally closer to 1000 words?

In addition, I've made some formatting and typographic edits, including fixes some issues to the bibliography. Can you review and merge? https://github.com/openmsr/CAD_to_OpenMC/pull/154

I also found an issue with the first sentence of the Statement of Need; the sentence does not seem complete, or at minimum needs editing for clarity. Can you address this?

Please let me know when you've made these edits; there is no need to archive the repo again, if the software does not change.
