### 2025-06-13 #7710 @ebknudsen
@kyleniemeyer Paper should now by my checks be somewhat below 2000 words. I don't think I can get it lower without removing things the review process put in. I completely removed the section about molten salt reactors. This will be included in the docs instead.
